requirements.txt
coc.py==1.0.4
discord-py-slash-command==1.0.8.5
discord.py==1.6.0
heroku3==4.2.3
psycopg2==2.8.6
python-dotenv==0.15.0
